feat(cxl-ui): rename chapter mixin to chapter navigation mixin
