+++ packages/cxl-ui/scss/cxl-jw-player/cxl-jw-player-chapter-navigation.scss
