+++ packages/cxl-ui/src/components/cxl-jw-player/index.js
@@ -1,6 +1,6 @@
 import { LitElement } from 'lit';
 import { customElement } from 'lit/decorators.js';
-import { BaseMixin, TranscriptMixin, ChapterMixin, SavePositionMixin } from './mixins';
+import { BaseMixin, TranscriptMixin, ChapterNavigationMixin, SavePositionMixin } from './mixins';
 import style from '../../styles/cxl-jw-player/cxl-jw-player-css';
 import shadowStyle from '../../styles/cxl-jw-player/cxl-jw-player-shadow-css';
 import { mixin } from './utility';
@@ -10,7 +10,7 @@ import { template } from './index.html';
 export class CXLJWPlayerElement extends mixin(LitElement, [
   BaseMixin,
   TranscriptMixin,
-  ChapterMixin,
+  ChapterNavigationMixin,
   SavePositionMixin,
 ]) {
   config = {
