+++ packages/cxl-ui/src/components/cxl-jw-player/mixins/index.js
@@ -1,4 +1,4 @@
 export { BaseMixin } from './BaseMixin';
 export { TranscriptMixin } from './TranscriptMixin';
-export { ChapterMixin } from './chapter';
+export { ChapterNavigationMixin } from './chapter-navigation';
 export { SavePositionMixin } from './SavePositionMixin';
